% Generated by roxygen2 (4.0.2): do not edit by hand
\name{psdrf_AgregArbres}
\alias{psdrf_AgregArbres}
\title{Aggrégation des arbres à l'échelle de la placette.}
\usage{
psdrf_AgregArbres()
}
\value{
La fonction construit les tables suivantes. Elles sont enregistrées dans le dossier Tables

psdrfPlaBMSlineV (volume de bois mort au sol par placettes, classe de diamètre, stadeD et stadeE),
psdrfPlaBMSlineVEss, psdrfPlaBMSlineVClasse, psdrfPlaBMSlineVStadeD,
psdrfPlaBMSlineVStadeE, psdrfPlaBMSsup30V, psdrfPlaBMSsup30VEss, psdrfPlaBMSsup30VClasse,
psdrfPlaBMSsup30VStadeD, psdrfPlaBMSsup30VStadeE, psdrfPlaBoisMT,
psdrfPlaBMT, psdrfPlaBMTStadeD, psdrfPlaBMTStadeE, psdrfPla,
psdrfPlaArbresV, psdrfPlaArbresVEss, psdrfPlaArbresVEssReg,
psdrfPlaArbresVClasse, psdrfPlaArbresVCat, psdrfPlaArbresVClasseEssReg,
psdrfPlaBMSinf, psdrfPlaBMSsup, psdrfPlaBMPinf, psdrfPlaBMPsup,psdrfPlaBM,
psdrfPlaRege,
}
\description{
Sous-programme permettant d'aggréger toutes les variables à l'échelle de la placette.
Il fournit de nombreux tableaux croisés.
}
\author{
Bruciamacchie Max
}

