% Generated by roxygen2 (4.0.2): do not edit by hand
\name{psdrf_AgregPlacettes}
\alias{psdrf_AgregPlacettes}
\title{Aggrégation des placettes à une échelle choisie par l'opérateur.}
\usage{
psdrf_AgregPlacettes()
}
\value{
La fonction construit les tables suivantes. Elles sont enregistrées dans le dossier Tables

psdrfNGVdisp, psdrfNGVstrate, psdrfNGVcc,
psdrfNGVdispEss, psdrfNGVdispEssReg, psdrfNGVdispCat,
psdrfHistRB, psdrfHistStrate, psdrfHistRBEssReg,
psdrfBMRB, psdrfBMRBstadeD, psdrfBMRBstadeE
}
\description{
Sous-programme permettant de calculer les moyennes, coefficient de variations et erreurs relatives par groupes de placettes
tels que choisis dans la feuille Placettes.
}
\author{
Bruciamacchie Max
}

