% Generated by roxygen2 (4.0.2): do not edit by hand
\name{psdrf_Calculs}
\alias{psdrf_Calculs}
\title{Calcul des variables dendrométriques et écologiques.}
\usage{
psdrf_Calculs()
}
\value{
La fonction construit les tables suivantes.

Arbres, BMP, BMSline, BMSsup30, Rege

Elles sont enregistrées dans le dossier Tables sous le nom : psdrfTablesBrutes.RData.
}
\description{
Cette fonction commence par calculer le poids de chaque arbre,
puis pour chacun d'eux, calcule les variables dendrométriques et écologiques ramenées à l'hectare.
}
\author{
Bruciamacchie Max
}

