% Generated by roxygen2 (4.0.2): do not edit by hand
\name{psdrf_ClasseurRem}
\alias{psdrf_ClasseurRem}
\title{Préparation du fichier de remesure.}
\usage{
psdrf_ClasseurRem()
}
\value{
La fonction construit un classeur Excel. Il est enregistré dans le dossier out
sous le nom : Remesures_NomDisp.xlsx

Exemple : Remesures_RB Grands Monts.xlsx.
}
\description{
Création d'un classeur destiné à faciliter la saisie de la prochaine campagne de mesure.
Il permettra également d'ajouter les nouvelles données dans les archives de la forêt.
Une boîte de dialogue permet de sélectionner la réserve souhaitée.
}
\author{
Bruciamacchie Max
}

