% Generated by roxygen2 (4.0.2): do not edit by hand
\name{psdrf_Codes}
\alias{psdrf_Codes}
\title{Création des tables de code.}
\usage{
psdrf_Codes()
}
\value{
La fonction construit les tables suivantes.

CodeDurete, CodeEcologie, CodeEcorce, CodeEssence, CodeTypoArbres, Communes, EssReg,
Reserves, Tiers, Tarifs,

Elles sont enregistrées dans le dossier Tables sous le nom : psdrfCodes.Rdata.
}
\description{
Creation d'une archive RData a partir du classeur Excel PsdrfListes.xlsx.
}
\author{
Bruciamacchie Max
}

