% Generated by roxygen2 (4.0.2): do not edit by hand
\name{psdrf_EditCarnet}
\alias{psdrf_EditCarnet}
\title{Edition du carnet de chaque dispositif.}
\usage{
psdrf_EditCarnet(format = 0)
}
\arguments{
\item{format}{paramètre permettant de fixer le format du fichier en sortie.}
}
\value{
Edition d'un rapport par dispositif
}
\description{
La fonction permet de sélectionner un ou plusieurs dispositifs et d'éditer un rapport aux formats html, pdf ou doc.
}
\details{
Le format de sortie sera  \cr
du .html si le paramètre vaut 0 (valeur par défaut),  \cr
du .pdf si le format est égal à 1,  \cr
du .doc si format est égal à 2
}
\author{
Bruciamacchie Max
}

