% Generated by roxygen2 (4.0.2): do not edit by hand
\name{psdrf_ShapesPlac}
\alias{psdrf_ShapesPlac}
\title{Jonction des tables attributaires avec le shape placettes.}
\usage{
psdrf_ShapesPlac()
}
\arguments{
\item{shp}{= shape des placettes avec comme système de projection RGF93/lambert-93.}
}
\value{
La fonction construit le shape PsdrfPlacettes.shp dans le dossier SIG/Vecteurs.
}
\description{
La fonction récupère les data.frames construits à l'échelle des placettes
pour faire la jonction des données dendrométriques (psdrfPla), de la composition en essence
(psdrfPlaArbresVEssReg) et des données de bois mort (psdrfPlaBM).
}
\author{
Bruciamacchie Max
}

