% Generated by roxygen2 (4.0.2): do not edit by hand
\name{psdrf_Tables2Xls}
\alias{psdrf_Tables2Xls}
\title{Export des tables elaborees au format xlsx.}
\usage{
psdrf_Tables2Xls()
}
\value{
La fonction exporte dans un classeur au format .xlsx les tables suivantes :

psdrfDendroCC, psdrfDendroReserve, psdrfDendroReserveCat, psdrfDendroReserveEss,
psdrfDendroReserveEssReg, psdrfDendroStrate

Le classeur est enregistré dans le dossier Résultats/Out sous le nom : TabPlacettes.xlsx.
}
\description{
Sauvegarde des tables élaborées dans un classeur Excel au format xlsx, dans un dossier Out.
}
\author{
Bruciamacchie Max
}

