% Generated by roxygen2 (4.0.2): do not edit by hand
\name{psdrf_Verif}
\alias{psdrf_Verif}
\title{Vérification des données brutes de chaque dispositif.}
\usage{
psdrf_Verif(repdata, modif = F)
}
\arguments{
\item{repdata}{= répertoire contenant les données}

\item{modif}{= argument permettant de modifier le format du rapport. Une boîte de dialogue permet de choisir
le format de sortie parmi les 3 suivants : .html, .pdf, .doc.}
}
\value{
Edition d'un rapport par dispositif.
}
\description{
La fonction permet de sélectionner un ou plusieurs dispositifs et d'éditer un rapport aux formats html, pdf ou doc.
}
\author{
Bruciamacchie Max
}

