% Generated by roxygen2 (4.0.2): do not edit by hand
\name{psdrf_Xls2Rdata}
\alias{psdrf_Xls2Rdata}
\title{Conversion des données brutes de la base PSDRF en 2 archives RData.}
\usage{
psdrf_Xls2Rdata(repdata)
}
\arguments{
\item{repdata}{= répertoire contenant les données}
}
\value{
La fonction exporte dans le dossier Tables l'archive suivante :

psdrfDonneesBrutes.Rdata
}
\description{
Import des classeurs Excel de chaque réserve et sauvegarde dans l'archive
psdrfDonneesBrutes.Rdata.

Import des différentes tables de codification et sauvegarde dans l'archive psdrfCodes.Rdata.
}
\author{
Bruciamacchie Max
}

